remove dasboards page
