+++ apps/frontend/app/(website)/campaigns/[campaignId]/components/DynamicLineChart.tsx
