+++ apps/frontend/app/(website)/campaigns/[campaignId]/mockData.tsx
