+++ apps/frontend/app/(website)/campaigns/[campaignId]/mocks/card-revenue.json
