+++ apps/frontend/app/(website)/campaigns/[campaignId]/mocks/chart-multiple.json
