+++ apps/frontend/app/(website)/campaigns/[campaignId]/mocks/chart-netlify.json
