+++ apps/frontend/app/(website)/campaigns/[campaignId]/types.ts
