+++ apps/frontend/app/(website)/campaigns/[campaignId]/widgets/CustomChart.tsx
@@ -1,7 +1,7 @@
 import {
   ChartTile,
   type ChartTileProps,
-} from "@/app/(website)/campaigns/[campaignId]/view/[viewId]/components/ChartTile";
+} from "@/app/(website)/campaigns/[campaignId]/components/ChartTile";
 
 export const CustomChart = ({
   data,
